LayoutTests/inspector/console-dirxml-expected.txt

CONSOLE MESSAGE: line 8: [object HTMLDocument]
Tests that console logging dumps proper messages.

console-dirxml.html:8Document
